+++ src/main/scala/org/mmisw/cf2rdf/OrrNvsMapper.scala
@@ -6,11 +6,15 @@ import java.nio.charset.StandardCharsets
 import org.apache.jena.ontology.OntModelSpec
 import org.apache.jena.rdf.model.{ModelFactory, Resource}
 import org.apache.jena.vocabulary.SKOS
+import org.mmisw.orr.ont.vocabulary.{Omv, OmvMmi}
 
 import scala.collection.JavaConversions._
 
 
-class OrrNvsMapper(nvsFilename: String) {
+class OrrNvsMapper(nvsFilename: String,
+                   versionNumberOpt: Option[String],
+                   lastModifiedOpt: Option[String]
+                  ) {
 
   private val mapNamespace   = "http://mmisw.org/ont/mmi/cfonmap/"
   private val orrCfNamespace = "http://mmisw.org/ont/cf/parameter/"
@@ -23,6 +27,39 @@ class OrrNvsMapper(nvsFilename: String) {
   model.setNsPrefix("nvs", nvsCfNamespace)
 
   model.setNsPrefix("skos", SKOS.uri)
+  model.setNsPrefix("omv", Omv.NS)
+  model.setNsPrefix("omvm", OmvMmi.NS)
+
+  private val ontology = model.createOntology("http://mmisw.org/ont/mmi/cfonmap")
+
+  ontology.addProperty(Omv.name, "ORR-NVS CF standard name mapping" +
+    (if (versionNumberOpt.isDefined) s" (v.${versionNumberOpt.get})" else ""))
+
+  ontology.addProperty(Omv.description,
+    "Uses skos:exactMatch to link the IRIs of the CF standard names between the" +
+      " RDF versions at the MMI ORR and NERC NVS repositories.")
+
+  ontology.addProperty(Omv.hasCreator, "MMI")
+  ontology.addProperty(OmvMmi.hasContentCreator, "MMI")
+
+  ontology.addProperty(Omv.keywords, {
+    List(
+      "NetCDF", "CF", "Climate and Forecast", "self-describing", "standard names", "Canonical Units"
+    ).mkString(", ")
+  })
+
+  ontology.addProperty(Omv.documentation, "https://github.com/mmisw/cf2rdf")
+  ontology.addProperty(Omv.reference,     "https://github.com/mmisw/cf2rdf")
+
+  ontology.addProperty(Omv.acronym, "cfontmap")
+
+  lastModifiedOpt foreach { lm ⇒
+    ontology.addProperty(OmvMmi.origVocLastModified, lm)
+    ontology.addProperty(Omv.creationDate, lm)
+  }
+
+  ontology.addProperty(OmvMmi.hasResourceType, "http://mmisw.org/ont/mmi/resourcetype/parameter")
+  ontology.addProperty(OmvMmi.origMaintainerCode, "mmi")
 
   private val map = loadNvs()